Do not install `frp`, `chisel`, `websocat`, `just` & `desed` by default

+++ getting_start.sh
@@ -162,10 +162,20 @@ else
   rm '/usr/local/bin/go-shadowsocks2' '/usr/local/bin/sslocal' '/usr/local/bin/ssmanager' '/usr/local/bin/ssserver' '/usr/local/bin/ssurl'
 fi
 
-curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/frpc" '/usr/local/bin/frpc'
-curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/frps" '/usr/local/bin/frps'
+# shellcheck disable=SC2154
+if [[ x"$(echo "${install_frp:=no}" | cut -c1)" = x'y' ]]; then
+  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/frpc" '/usr/local/bin/frpc'
+  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/frps" '/usr/local/bin/frps'
+else
+  rm '/usr/local/bin/frpc' '/usr/local/bin/frps'
+fi
 
-curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/chisel" '/usr/local/bin/chisel'
+# shellcheck disable=SC2154
+if [[ x"$(echo "${install_chisel:=no}" | cut -c1)" = x'y' ]]; then
+  curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/chisel" '/usr/local/bin/chisel'
+else
+  rm '/usr/local/bin/chisel'
+fi
 
 curl_to_dest "https://github.com/IceCodeNew/go-collection/releases/latest/download/got" '/usr/local/bin/got'
 
@@ -180,13 +190,26 @@ dog_latest_tag_name="$(curl -sSL -H 'Accept: application/vnd.github.v3+json' \
 curl "https://github.com/ogham/dog/releases/download/${dog_latest_tag_name}/dog-${dog_latest_tag_name}-x86_64-unknown-linux-gnu.zip" | bsdtar -xf- -P -C /usr/local
 curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/dog" '/usr/local/bin/dog'
 
-curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/websocat" '/usr/local/bin/websocat'
+# shellcheck disable=SC2154
+if [[ x"$(echo "${install_websocat:=no}" | cut -c1)" = x'y' ]]; then
+  curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/websocat" '/usr/local/bin/websocat'
+else
+  rm '/usr/local/bin/websocat'
+fi
 
-# curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/just" '/usr/local/bin/just'
-rm '/usr/local/bin/just'
+# shellcheck disable=SC2154
+if [[ x"$(echo "${install_just:=no}" | cut -c1)" = x'y' ]]; then
+  curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/just" '/usr/local/bin/just'
+else
+  rm '/usr/local/bin/just'
+fi
 
-# curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/desed" '/usr/local/bin/desed'
-rm '/usr/local/bin/desed'
+# shellcheck disable=SC2154
+if [[ x"$(echo "${install_desed:=no}" | cut -c1)" = x'y' ]]; then
+  curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/desed" '/usr/local/bin/desed'
+else
+  rm '/usr/local/bin/desed'
+fi
 
 curl_to_dest "https://github.com/IceCodeNew/rust-collection/releases/latest/download/fnm" '/usr/local/bin/fnm'
